Test/BufferWriter: Remove StringContainer; Improve documentation
This removes the unused StringContainer structure in Test.
This was used to test wire types; however, we know they're ok now.

The test program is essentially just here for me to try out ideas,
however it should still be OK code-quality wise.

This commit also changes the behaviour of BufferWriter::Release() a bit,
by making it only copy how many bytes were written rather than the size
of the backing vector. This should make it possible to write a message,
only allocating the buffer once and only reallocating for a huge data
payload.